+++ README.md
@@ -1,0 +1,4 @@
+# Researching design patterns using TypeScript
+
+### Changelog:
+- 10dec2023: adapter-v2.0: `src/design-patterns/Adapter`
