+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "test": "jest",
+    "test": "jest --watch",
     "lint": "eslint . --ext .ts"
   },
   "author": "",
